Implemented AppDataAccessKeeper to GetConfig and WriteConfig

+++ D_API/Controllers/AppDataHost.cs
@@ -32,7 +32,7 @@ namespace D_API.Controllers
         public async Task<IActionResult> GetConfig(string appname)
         {
             var s = Directories.InAppDataHost(appname);
-            return System.IO.File.Exists(s) ? Ok(await ReadFile(s)) : NotFound();
+            return System.IO.File.Exists(s) ? await Keeper.CheckAccess(User.Identity!.Name!, s) ? Ok(await ReadFile(s)) : Forbid() : NotFound();
         }
 
         [HttpPost("config/{appname}")]
@@ -42,6 +42,12 @@ namespace D_API.Controllers
         public async Task<IActionResult> WriteConfig(string appname, [FromBody]string body, bool ow)
         {
             var file = Directories.InAppDataHost(appname);
+
+            if (!await Keeper.CheckAccess(User.Identity!.Name!, file))
+                return Forbid();
+
+            await Keeper.NewFile(User.Identity!.Name!, file);
+
             if (!ow && System.IO.File.Exists(file))
                 return Unauthorized("File already exists, cannot assume overwrite. If you wish to overwrite it, please use \"config/{appname}/true\"");
             await WriteFile(file, body);
